renamed files according to autocat current convetion.  fixed some of my code/comment formatting (used rufus). centralized some url logic in opds.
